+++ custom/icds_reports/utils/monolith/agg_awc_daily.py
@@ -2,11 +2,11 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 import six
-
 from six.moves import map
 
 from corehq.util.python_compatibility import soft_assert_type_text
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, transform_day_to_month
+from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class AggAwcDailyAggregationHelper(BaseICDSAggregationHelper):
