+++ custom/icds_reports/utils/monolith/agg_child_health.py
@@ -2,13 +2,11 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 import six
-
-from corehq.apps.userreports.models import StaticDataSourceConfiguration, get_datasource_config
-from corehq.apps.userreports.util import get_table_name
 from six.moves import map
 
 from corehq.util.python_compatibility import soft_assert_type_text
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, transform_day_to_month
+from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class AggChildHealthAggregationHelper(BaseICDSAggregationHelper):
