+++ custom/icds_reports/utils/monolith/agg_ls_data.py
@@ -2,12 +2,13 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 from dateutil.relativedelta import relativedelta
-from custom.icds_reports.const import AGG_LS_VHND_TABLE, AGG_LS_AWC_VISIT_TABLE, AGG_LS_BENEFICIARY_TABLE
+from six.moves import range
+
 from corehq.apps.userreports.models import StaticDataSourceConfiguration, get_datasource_config
 from corehq.apps.userreports.util import get_table_name
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, \
-    transform_day_to_month, month_formatter
-from six.moves import range
+from custom.icds_reports.const import AGG_LS_VHND_TABLE, AGG_LS_AWC_VISIT_TABLE, AGG_LS_BENEFICIARY_TABLE
+from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month, month_formatter
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class AggLsHelper(BaseICDSAggregationHelper):
