+++ custom/icds_reports/utils/monolith/awc_location.py
@@ -1,11 +1,12 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
+from six.moves import map
+
 from corehq.apps.userreports.models import StaticDataSourceConfiguration, get_datasource_config
 from corehq.apps.userreports.util import get_table_name
 from custom.icds_reports.const import AWC_LOCATION_TABLE_ID, AWW_USER_TABLE_ID
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper
-from six.moves import map
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class LocationAggregationHelper(BaseICDSAggregationHelper):
