+++ custom/icds_reports/utils/monolith/aww_incentive.py
@@ -3,12 +3,12 @@ from __future__ import unicode_literals
 
 from corehq.apps.userreports.models import StaticDataSourceConfiguration, get_datasource_config
 from corehq.apps.userreports.util import get_table_name
-
 from custom.icds_reports.const import (
     AWW_INCENTIVE_TABLE,
     AGG_CCS_RECORD_CF_TABLE
 )
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, month_formatter
+from custom.icds_reports.utils.aggregation_helpers import month_formatter
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class AwwIncentiveAggregationHelper(BaseICDSAggregationHelper):
