+++ custom/icds_reports/utils/monolith/ccs_record_monthly.py
@@ -11,7 +11,8 @@ from custom.icds_reports.const import (
     AGG_CCS_RECORD_THR_TABLE,
     AGG_CCS_RECORD_DELIVERY_TABLE,
     AGG_CCS_RECORD_CF_TABLE)
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, transform_day_to_month
+from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class CcsRecordMonthlyAggregationHelper(BaseICDSAggregationHelper):
