+++ custom/icds_reports/utils/monolith/child_health_monthly.py
@@ -12,8 +12,9 @@ from custom.icds_reports.const import (
     AGG_DAILY_FEEDING_TABLE,
     AGG_GROWTH_MONITORING_TABLE,
 )
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, transform_day_to_month, \
+from custom.icds_reports.utils.aggregation_helpers import transform_day_to_month, \
     month_formatter
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class ChildHealthMonthlyAggregationHelper(BaseICDSAggregationHelper):
