+++ custom/icds_reports/utils/monolith/daily_attendance.py
@@ -4,8 +4,9 @@ from __future__ import unicode_literals
 from corehq.apps.userreports.models import StaticDataSourceConfiguration, get_datasource_config
 from corehq.apps.userreports.util import get_table_name
 from custom.icds_reports.const import DAILY_FEEDING_TABLE_ID
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, date_to_string, \
+from custom.icds_reports.utils.aggregation_helpers import date_to_string, \
     transform_day_to_month
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class DailyAttendanceAggregationHelper(BaseICDSAggregationHelper):
