+++ custom/icds_reports/utils/monolith/ls_vhnd_form.py
@@ -2,8 +2,10 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 from dateutil.relativedelta import relativedelta
+
 from custom.icds_reports.const import AGG_LS_VHND_TABLE
-from custom.icds_reports.utils.aggregation_helpers import BaseICDSAggregationHelper, month_formatter
+from custom.icds_reports.utils.aggregation_helpers import month_formatter
+from custom.icds_reports.utils.monolith import BaseICDSAggregationHelper
 
 
 class LSVhndFormAggHelper(BaseICDSAggregationHelper):
